+++ tinytuya/__init__.py
@@ -97,3 +97,8 @@ from .OutletDevice import OutletDevice
 from .CoverDevice import CoverDevice
 from .BulbDevice import BulbDevice
 from .Cloud import Cloud
+
+from .crypto_helper import *
+from .message_helper import *
+from .exceptions import *
+from .error_helper import *
