+++ tinytuya/error_helper.py
@@ -1,0 +1,58 @@
+# TinyTuya Module
+# -*- coding: utf-8 -*-
+
+import json
+import logging
+
+log = logging.getLogger(__name__)
+
+
+# TinyTuya Error Response Codes
+ERR_JSON = 900
+ERR_CONNECT = 901
+ERR_TIMEOUT = 902
+ERR_RANGE = 903
+ERR_PAYLOAD = 904
+ERR_OFFLINE = 905
+ERR_STATE = 906
+ERR_FUNCTION = 907
+ERR_DEVTYPE = 908
+ERR_CLOUDKEY = 909
+ERR_CLOUDRESP = 910
+ERR_CLOUDTOKEN = 911
+ERR_PARAMS = 912
+ERR_CLOUD = 913
+ERR_KEY_OR_VER = 914
+
+error_codes = {
+    ERR_JSON: "Invalid JSON Response from Device",
+    ERR_CONNECT: "Network Error: Unable to Connect",
+    ERR_TIMEOUT: "Timeout Waiting for Device",
+    ERR_RANGE: "Specified Value Out of Range",
+    ERR_PAYLOAD: "Unexpected Payload from Device",
+    ERR_OFFLINE: "Network Error: Device Unreachable",
+    ERR_STATE: "Device in Unknown State",
+    ERR_FUNCTION: "Function Not Supported by Device",
+    ERR_DEVTYPE: "Device22 Detected: Retry Command",
+    ERR_CLOUDKEY: "Missing Tuya Cloud Key and Secret",
+    ERR_CLOUDRESP: "Invalid JSON Response from Cloud",
+    ERR_CLOUDTOKEN: "Unable to Get Cloud Token",
+    ERR_PARAMS: "Missing Function Parameters",
+    ERR_CLOUD: "Error Response from Tuya Cloud",
+    ERR_KEY_OR_VER: "Check device key or version",
+    None: "Unknown Error",
+}
+
+
+def error_json(number=None, payload=None):
+    """Return error details in JSON"""
+    try:
+        spayload = json.dumps(payload)
+        # spayload = payload.replace('\"','').replace('\'','')
+    except:
+        spayload = '""'
+
+    vals = (error_codes[number], str(number), spayload)
+    log.debug("ERROR %s - %s - payload: %s", *vals)
+
+    return json.loads('{ "Error":"%s", "Err":"%s", "Payload":%s }' % vals)
